+++ NEWS.md
@@ -15,6 +15,10 @@ Zelig version 5.1-2
 data or bootstraps an error is no longer returned. Instead a list of the 
 models' coefficients, standard errors, and p-values is returned. #268
 
+-   `summary` for `logt` models now accepts the argument `odds_ratios`. When
+`TRUE` odds ratio estimates are returned rather than coefficient estimates.
+Thanks to Adam Obeng. PR/#270.
+
 ## Minor changes and bug fixes
 
 -   Resolved a bug where `weights` were not being passed to `svydesign`